requirements.txt
django==2.2
djangorestframework==3.9.2
#django-rest-swagger
wheel
django-filter==0.15.3
